Update order fix for empty orders

+++ update_order.php
@@ -18,7 +18,7 @@ if (isset($_POST['orderID'])) {
     foreach($quanty as $itemy => $quantidy) {
       // itemPrice for each item
       $iip = $_POST['itemPrice'][$itemy];
-    if (isset($_POST['oldQuant'])) {
+    if (isset($_POST['oldQuant'][$itemy] && $_POST['oldQuant'][$itemy] > 0)) {
       if ($quantidy != $_POST['oldQuant'][$itemy]) {
         if($quantidy == 0) {
           $deleteQuery = "DELETE from CONTAINS where orderID = $_POST[orderID] and itemID = $itemy";
@@ -98,7 +98,6 @@ if (isset($_POST['orderID'])) {
     <?php
       require_once '/home/hipt3660/config/mysql_config.php';
       $sql = "select * from ORDERS where orderID = {$_GET['orderID']}";
-      $sqlc = "select * from CONTAINS where orderID = {$_GET['orderID']}";
       
       try {
         $result = $conn->query($sql);
@@ -106,63 +105,50 @@ if (isset($_POST['orderID'])) {
         echo $e->getMessage();
       }
 
-      try {
-        $resultc = $conn->query($sqlc);
-      } catch (Exception $e) {
-        echo $e->getMessage();
-      }
-
       if ($result->num_rows != 0) {
         $order = $result->fetch_assoc();
-        $contains = $resultc->fetch_assoc();
-        $sqliname = "SELECT itemName FROM MENU WHERE itemID = {$contains['itemID']}";
-        try {
-            $resiname = $conn->query($sqliname);
-          } catch (Exception $e) {
-            echo $e->getMessage();
-          }
-        $itemname = $resiname->fetch_assoc();
-            echo "<h2>Update Information for Order {$order['orderID']}</h2>";
+        echo "<h2>Update Information for Order {$order['orderID']}</h2>";
         echo "<form action='' method='post'>";
         echo "Status: <input type=text name='status' value='{$order['status']}' size=20><br><br>";
         echo "Employee ID: <input type=text name='empID' value={$order['empID']} size=6><br><br>";
         echo '<table> <tr>
-    <td> Name </td> 
-    <td> Price </td> 
-    <td> Vegan </td> 
-    <td> Stock </td> 
-    <td> Quantity </td>
-    </tr>';
+        <td> Name </td> 
+        <td> Price </td> 
+        <td> Vegan </td> 
+        <td> Stock </td> 
+        <td> Quantity </td>
+        </tr>';
         $sql = "SELECT * FROM MENU";
         $result = $conn->query($sql);
-      if($result->num_rows > 0) {
-        while($row = $result->fetch_assoc()) {
-            $field2name = $row["itemName"];
-            $field3name = $row["itemPrice"];
-            $field4name = $row["isVegan"];
-            $field5name = $row["stock"];
+        if($result->num_rows > 0) {
+          while($row = $result->fetch_assoc()) {
+              $field2name = $row["itemName"];
+              $field3name = $row["itemPrice"];
+              $field4name = $row["isVegan"];
+              $field5name = $row["stock"];
 
-            $vegan = ($field4name == 0 ? 'No' : 'Yes');
+              $vegan = ($field4name == 0 ? 'No' : 'Yes');
 
-            // display quantity for item in that row where itemID = itemID
-            $sqlq = "SELECT quantity FROM CONTAINS WHERE itemID = {$row['itemID']} AND orderID = {$_GET['orderID']}";
-            try {
-              $resultq = $conn->query($sqlq);
-            } catch (Exception $e) {
-              echo $e->getMessage();
-            }
-            $rowq = $resultq->fetch_assoc();
+              // display quantity for item in that row where itemID = itemID
+              $sqlq = "SELECT quantity FROM CONTAINS WHERE itemID = {$row['itemID']} AND orderID = {$_GET['orderID']}";
+              try {
+                $resultq = $conn->query($sqlq);
+              } catch (Exception $e) {
+                echo $e->getMessage();
+              }
+              $rowq = $resultq->fetch_assoc();
 
-            echo '<tr> 
-                      <td>'. $field2name.'</td> 
-                      <td>'. $field3name.'</td> 
-                      <td>'. $vegan. '</td> 
-                      <td>'. $field5name."</td> 
-                      <td><input type='number' name='add[$row[itemID]]' min='0' max='$field5name' value='$rowq[quantity]'></td>
-                      <input type='hidden' name='oldQuant[$row[itemID]]' value='$rowq[quantity]'>
-                      <input type='hidden' name='itemPrice[$row[itemID]]' value='$row[itemPrice]'>
-                  </tr>";
-        }}
+              echo '<tr> 
+                        <td>'. $field2name.'</td> 
+                        <td>'. $field3name.'</td> 
+                        <td>'. $vegan. '</td> 
+                        <td>'. $field5name."</td> 
+                        <td><input type='number' name='add[$row[itemID]]' min='0' max='$field5name' value='$rowq[quantity]'></td>
+                        <input type='hidden' name='oldQuant[$row[itemID]]' value='$rowq[quantity]'>
+                        <input type='hidden' name='itemPrice[$row[itemID]]' value='$row[itemPrice]'>
+                    </tr>";
+          }
+        }
         echo "Order Completed?: ";
         if ($order['isComplete']) {
           echo "<input type=checkbox name='isComplete' checked><br><br>";
